fix(core): fixed code smells of sonar
