+++ libs/core/src/test/java/com/sourcefuse/jarc/core/CoreApplicationTests.java
@@ -1,5 +1,7 @@
 package com.sourcefuse.jarc.core;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import com.sourcefuse.jarc.core.repositories.SoftDeletesRepositoryImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,5 +12,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 class CoreApplicationTests {
 
   @Test
-  void contextLoads() {}
+  void contextLoads() {
+    assertTrue(true);
+  }
 }
